requirements.txt
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.1
djangorestframework==3.13.1
drf-spectacular==0.21.1
flake8==4.0.1
idna==3.3
importlib-resources==5.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
openapi-codec==1.3.2
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytz==2021.3
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
sqlparse==0.4.2
tomli==1.2.3
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
zipp==3.7.0
